[chore] 컴포넌트 뷰 분리(SettingComponents)
