+++ AhnNyeong/AhnNyeong/UI/SettingViews/SettingNotiView.swift
@@ -76,107 +76,6 @@ struct SettingNotiView: View {
         .navigationBarTitleDisplayMode(.inline)
         .navigationBarItems(leading: BackButton(backBtnTitleType: .titleImage, backButtonTitle: "gearshape.fill"))
         .background(Color.white300)
-        
-    }
-}
-
-enum BackBtnTitleType {
-    case titleText
-    case titleImage
-}
-
-struct BackButton: View {
-    @Environment(\.dismiss) private var dismiss
-    let backBtnTitleType: BackBtnTitleType
-    let backButtonTitle: String
-    var body: some View {
-        HStack(spacing: 0) {
-            Image(systemName: "chevron.left")
-                .font(.system(size: 17, weight: .semibold))
-                .padding(.trailing, 4)
-            switch backBtnTitleType {
-            case .titleText:
-                Text(backButtonTitle)
-                    .font(.system(size: 17, weight: .regular))
-            case .titleImage:
-                Image(systemName: backButtonTitle)
-                    .font(.system(size: 17, weight: .semibold)) // need some check!
-            }
-        }
-        .foregroundColor(.coral500)
-        .onTapGesture {
-            dismiss()
-        }
-    }
-}
-
-struct SettingViewTitle: View {
-    let settingViewTitle: String
-    var body: some View {
-        VStack(alignment: .leading, spacing: 0) {
-            Text(settingViewTitle)
-                .font(.system(size: 28, weight: .bold))
-                .foregroundColor(.black400)
-                .padding(.horizontal, 16)
-                .padding(.top, 20)
-            
-            DividingRectangle(dividingType: .naviTitleDivider)
-                .padding(.top, 20)
-        }
-    }
-}
-
-struct ListTitle: View {
-    let listTitle: String
-    let listCaption: String
-    var body: some View {
-        VStack(alignment: .leading, spacing: 0) {
-            Text(listTitle)
-                .bold18Black400()
-            Text(listCaption)
-                .medium12Black300()
-        }
-    }
-}
-
-struct ToggleList: View {
-    let toggleTitle: String
-    @Binding var toggleIsOn: Bool
-    var body: some View {
-        Toggle(toggleTitle, isOn: $toggleIsOn)
-            .padding(.vertical, 6.5)
-            .padding(.horizontal, 16)
-    }
-}
-
-struct DatePickerList: View {
-    let datePickerTitle: String
-    @Binding var selection: Date
-    var body: some View {
-        DatePicker(datePickerTitle, selection: $selection, displayedComponents: .hourAndMinute)
-            .padding(.vertical, 6.5)
-            .padding(.horizontal, 16)
-    }
-}
-enum DividingType {
-    case naviTitleDivider
-    case listDivider
-}
-
-struct DividingRectangle: View {
-    let dividingType: DividingType
-
-    var body: some View {
-        switch dividingType {
-        case .naviTitleDivider:
-            return Rectangle()
-                .frame(height: 1)
-                .foregroundColor(.black75)
-        case .listDivider:
-            return Rectangle()
-                .frame(height: 0.5)
-                .foregroundColor(.black75)
-        }
     }
 }
